#91 - accepted parsing of commas after values in ddi / changed float to double in HIVE's ValType=Real test to accept larger values

+++ wasphive/HIVE.i.h
@@ -463,7 +463,7 @@ bool HIVE::validateValType(SchemaAdapter& schema_node, InputAdapter& input_node
             }
 
             if (ruleValue == "Real") {
-                float ftest;
+                double ftest;
                 iss >> std::noskipws >> ftest;
             }
